:truck: renaming files, migrate to Sensor.Community
