+++ Pipfile
@@ -1,0 +1,26 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+altgraph = "==0.16.1"
+certifi = "==2018.8.24"
+chardet = "==3.0.4"
+ecdsa = "==0.13"
+esptool = "==2.5.0"
+future = "==0.16.0"
+idna = "==2.7"
+macholib = "==1.11"
+netifaces = "==0.10.7"
+pefile = "==2018.8.8"
+pyaes = "==1.6.1"
+pyserial = "==3.4"
+requests = "==2.20.0"
+urllib3 = "==1.24.2"
+zeroconf = "==0.20.0"
+PyInstaller = "==3.4"
+PyQt5 = "==5.11.2"
+PyQt5-sip = "==4.19.19"
