+++ airrohr-flasher.spec
@@ -15,12 +15,12 @@ except Exception as exc:
 
 builddate = datetime.datetime.now().strftime('%Y%m%d')
 
-with open('luftdatentool/_buildid.py', 'w') as fd:
-    fd.write('''# This file is autogenerated in luftdaten-tool.spec file
+with open('airrohrFlasher/_buildid.py', 'w') as fd:
+    fd.write('''# This file is autogenerated in airrohr-flasher.spec file
 commit = "{commit}"
 builddate = "{builddate}"'''.format(commit=commit, builddate=builddate))
 
-a = Analysis(['luftdaten-tool.py'],
+a = Analysis(['airrohr-flasher.py'],
              pathex=['.'],
              binaries=[],
              datas=[('assets/', './assets'), ('i18n/*.qm', './i18n')],
@@ -38,7 +38,7 @@ exe = EXE(pyz,
           a.binaries,
           a.zipfiles,
           a.datas,
-          name='luftdaten-tool',
+          name='airrohr-flasher',
           debug=False,
           strip=False,
           upx=True,
@@ -48,6 +48,6 @@ exe = EXE(pyz,
 
 # This is used on MacOS only
 app = BUNDLE(exe,
-             name='Luftdaten.info Flashing Tool.app',
+             name='Sensor.Community Airrohr Flasher.app',
              icon='assets/logo.icns',
              bundle_identifier=None)
