+++ airrohrFlasher/qtvariant.py
