+++ requirements.txt
@@ -11,7 +11,7 @@ pefile==2018.8.8
 pyaes==1.6.1
 PyInstaller==3.4
 PyQt5==5.11.2
-PyQt5-sip==4.19.12
+PyQt5-sip==4.19.19
 pyserial==3.4
 requests==2.20.0
 urllib3==1.24.2
